data-collection/requirements.txt
pytest>=7.4.0
pytest-mock>=3.11.1
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
psycopg2-binary>=2.9.9
requests>=2.31.0
beautifulsoup4>=4.12.2
python-dotenv>=1.0.0
zeep>=4.2.1



